==> notebooks/06_split_data.ipynb <==
import pandas as pd
import os

# Paths
PROCESSED_PATH = "../data/processed/tabular/"
INPUT_FILE = os.path.join(PROCESSED_PATH, "supply_chain_features.csv")
OUTPUT_TRAIN = os.path.join(PROCESSED_PATH, "train.csv")
OUTPUT_VAL = os.path.join(PROCESSED_PATH, "val.csv")
OUTPUT_TEST = os.path.join(PROCESSED_PATH, "test.csv")

# Load dataset
df = pd.read_csv(INPUT_FILE, parse_dates=["dispatch_date", "delivery_date"])
print("✅ Loaded enriched dataset:", df.shape)


n = len(df)
train_end = int(0.7 * n)
val_end = int(0.85 * n)

df_train = df.iloc[:train_end]
df_val = df.iloc[train_end:val_end]
df_test = df.iloc[val_end:]

print(f"Train: {df_train.shape}, Val: {df_val.shape}, Test: {df_test.shape}")

df_train.to_csv(OUTPUT_TRAIN, index=False)
df_val.to_csv(OUTPUT_VAL, index=False)
df_test.to_csv(OUTPUT_TEST, index=False)

print("💾 Saved train/val/test splits:")
print(f"- {OUTPUT_TRAIN}")
print(f"- {OUTPUT_VAL}")
print(f"- {OUTPUT_TEST}")
